Handle better what are the recent releases included in timeline.
Always include releases from notable companies.
Adjust  popularity thresholds for inclusion in timeline.
Force refreshes for GameEntries of new releases and refresh SteamData for the first few days of release.

This is now a frankenstein script that uses two different firestore libraries. Need to replace the old firestore lib in the backend.