Change spawned task instrumentation to info.
Events were not attaching to the lower priority spans.
This is a bug as they should find the closest info span to attach into. Added a todo on the layer.

In the meantime decided to reduce the amount of spans by using spawns as info and removing lower level spans.