Change the logging schema for extracting metrics in gcloud.
Rust enums are cool but their serialization to json through Valuable does not allow for easy metrics extraction in gcloud :/
Simplifying the structure that allows for simpler querying in logs without nesting of spans. All events are thrown flat in the request so order and semantic grouping is lost.

Known issues
- Steam events should merge similar to Firestore events.
- Latencies to other service requests are lost. Need to find a new representation.