Consolidate Steam scraping inside the main Steam API.
The scraping is significantly slower and should not count to API limits so it happens in parallel to the API calls.

This also breaks up the Steam retrieval in 3 flavours for increasing speed.
When resolving a GameDigest there are only 2 steam API calls needed.
Makes also the Steam scraping API private.